// migrations/snapshots/f8e0031817b0403c8ba8459bf482dea59f5f3d8236574d96b204873d9cfa49e6/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Organization: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'] | null;
      readonly slug: CodecTypes['pg/text@1']['output'];
      readonly logoUrl: CodecTypes['pg/text@1']['output'] | null;
      readonly description: CodecTypes['pg/text@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['output'];
    };
    readonly OrganizationMember: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly organizationId: CodecTypes['pg/int4@1']['output'];
      readonly userId: CodecTypes['pg/int4@1']['output'];
      readonly role: 'ADMIN' | 'MEMBER' | 'OWNER';
      readonly joinedAt: CodecTypes['pg/timestamptz-string@1']['output'];
    };
    readonly User: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly password: CodecTypes['pg/text@1']['output'];
      readonly firstName: CodecTypes['pg/text@1']['output'] | null;
      readonly lastName: CodecTypes['pg/text@1']['output'] | null;
      readonly role: 'admin' | 'manager' | 'user';
      readonly status: 'active' | 'inactive' | 'suspended';
      readonly refreshTokenHash: CodecTypes['pg/text@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Organization: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'] | null;
      readonly slug: CodecTypes['pg/text@1']['input'];
      readonly logoUrl: CodecTypes['pg/text@1']['input'] | null;
      readonly description: CodecTypes['pg/text@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['input'];
    };
    readonly OrganizationMember: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly organizationId: CodecTypes['pg/int4@1']['input'];
      readonly userId: CodecTypes['pg/int4@1']['input'];
      readonly role: 'ADMIN' | 'MEMBER' | 'OWNER';
      readonly joinedAt: CodecTypes['pg/timestamptz-string@1']['input'];
    };
    readonly User: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly password: CodecTypes['pg/text@1']['input'];
      readonly firstName: CodecTypes['pg/text@1']['input'] | null;
      readonly lastName: CodecTypes['pg/text@1']['input'] | null;
      readonly role: 'admin' | 'manager' | 'user';
      readonly status: 'active' | 'inactive' | 'suspended';
      readonly refreshTokenHash: CodecTypes['pg/text@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['input'];
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly organization: {
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly description: CodecTypes['pg/text@1']['output'] | null;
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly logoUrl: CodecTypes['pg/text@1']['output'] | null;
      readonly name: CodecTypes['pg/text@1']['output'] | null;
      readonly slug: CodecTypes['pg/text@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['output'];
    };
    readonly organizationMember: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly joinedAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly organizationId: CodecTypes['pg/int4@1']['output'];
      readonly role: 'ADMIN' | 'MEMBER' | 'OWNER';
      readonly userId: CodecTypes['pg/int4@1']['output'];
    };
    readonly user: {
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly firstName: CodecTypes['pg/text@1']['output'] | null;
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly lastName: CodecTypes['pg/text@1']['output'] | null;
      readonly password: CodecTypes['pg/text@1']['output'];
      readonly refreshTokenHash: CodecTypes['pg/text@1']['output'] | null;
      readonly role: 'admin' | 'manager' | 'user';
      readonly status: 'active' | 'inactive' | 'suspended';
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['output'];
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly organization: {
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly description: CodecTypes['pg/text@1']['input'] | null;
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly logoUrl: CodecTypes['pg/text@1']['input'] | null;
      readonly name: CodecTypes['pg/text@1']['input'] | null;
      readonly slug: CodecTypes['pg/text@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['input'];
    };
    readonly organizationMember: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly joinedAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly organizationId: CodecTypes['pg/int4@1']['input'];
      readonly role: 'ADMIN' | 'MEMBER' | 'OWNER';
      readonly userId: CodecTypes['pg/int4@1']['input'];
    };
    readonly user: {
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly firstName: CodecTypes['pg/text@1']['input'] | null;
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly lastName: CodecTypes['pg/text@1']['input'] | null;
      readonly password: CodecTypes['pg/text@1']['input'];
      readonly refreshTokenHash: CodecTypes['pg/text@1']['input'] | null;
      readonly role: 'admin' | 'manager' | 'user';
      readonly status: 'active' | 'inactive' | 'suspended';
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['input'];
    };
  };
};
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;